# perf: count(*) for append-only tables 