--- a/src/common/query/Cargo.toml
+++ b/src/common/query/Cargo.toml
@@ -12,7 +12,6 @@ workspace = true
 
 [dependencies]
 api.workspace = true
-arrow-schema.workspace = true
 async-trait.workspace = true
 bytes.workspace = true
 common-error.workspace = true
@@ -22,7 +21,6 @@ common-time.workspace = true
 datafusion.workspace = true
 datafusion-common.workspace = true
 datafusion-expr.workspace = true
-datafusion-physical-expr.workspace = true
 datatypes.workspace = true
 serde.workspace = true
 snafu.workspace = true