--- a/src/mito2/src/read/unordered_scan.rs
+++ b/src/mito2/src/read/unordered_scan.rs
@@ -58,8 +58,12 @@ pub struct UnorderedScan {
 impl UnorderedScan {
     /// Creates a new [UnorderedScan].
     pub(crate) fn new(input: ScanInput) -> Self {
-        let properties =
-            ScannerProperties::new_with_partitions(input.parallelism.parallelism.max(1));
+        let parallelism = input.parallelism.parallelism.max(1);
+        let properties = ScannerProperties {
+            partitions: vec![vec![]; parallelism],
+            append_mode: input.append_mode,
+            total_rows: input.total_rows(),
+        };
         let stream_ctx = Arc::new(StreamContext::new(input));
 
         Self {
@@ -135,7 +139,11 @@ impl RegionScanner for UnorderedScan {
     }
 
     fn prepare(&mut self, ranges: Vec<Vec<PartitionRange>>) -> Result<(), BoxedError> {
-        self.properties = ScannerProperties::new(ranges);
+        self.properties = ScannerProperties::new(
+            ranges,
+            self.properties.append_mode,
+            self.properties.total_rows,
+        );
         Ok(())
     }