I hereby agree to the terms of the [GreptimeDB CLA](https://github.com/GreptimeTeam/.github/blob/main/CLA.md).

## Refer to a related PR or issue link (optional)
- closes #4516

## What's changed and what's your intention?

Optimize `count(*)` qeuries for tables with append mode enabled:

Test result on TSBS dataset (103680000 rows, append_mode=true): ~600x faster

```
# Before (approx. 4160ms)
mysql> select count(*) from benchmark.cpu;
+-----------+
| COUNT(*)  |
+-----------+
| 103680000 |
+-----------+
1 row in set (4.21 sec)

# After (approx. 7ms)
mysql> select count(*) from benchmark.cpu;
+-----------+
| COUNT(*)  |
+-----------+
| 103680000 |
+-----------+
1 row in set (0.01 sec)
```

## Checklist

- [ ] I have written the necessary rustdoc comments.
- [ ] I have added the necessary unit tests and integration tests.
- [ ] This PR requires documentation updates.